requirements.txt
apscheduler==3.11.2
asgiref==3.10.0
boto3==1.41.5
botocore==1.41.5
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
Django==5.2.7
django-apscheduler==0.7.0
django-environ==0.12.0
django-storages==1.14.6
django-template-maths==0.2.0
et-xmlfile==2.0.0
idna==3.11
jmespath==1.0.1
openpyxl==3.1.5
pillow==12.0.0
pytesseract==0.3.13
psycopg2-binary==2.9.11
pycparser==3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-http-client==3.3.7
requests==2.32.5
s3transfer==0.15.0
sendgrid==6.12.5
six==1.17.0
sqlparse==0.5.3
tzlocal==5.3.1
urllib3==2.5.0
